backend/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cloudinary==1.23.0
cryptography==3.3
distlib==0.3.1
dj3-cloudinary-storage==0.0.3
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-oauth-toolkit==1.3.3
django-rest-knox==4.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
idna==2.10
oauthlib==3.1.0
Pillow==8.0.1
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
virtualenv-clone==0.5.4
